backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.0
nltk==3.8.1
python-dotenv==1.0.0
requests==2.31.0